Use system setting to switch color schemes
Automatically switch between light and dark color schemes based on a
`prefers-color-scheme` media query. Tailwind is given
`var(--colors-scale-step)` as the value of its colors, and two new
utility classes (`.color-scheme-dark` and `.color-scheme-light`) can be
used to assign the variables. The plugin automatically applies those to
`:root` based on the media query.